extract ContainerHomeIpPortsPreview component
