+++ src/components/ContainerHomePreview.react.js
@@ -1,9 +1,9 @@
-import _ from 'underscore';
 import React from 'react/addons';
 import request from 'request';
 import shell from 'shell';
 import metrics from '../utils/MetricsUtil';
 import ContainerHomeWebPreview from './ContainerHomeWebPreview.react';
+import ContainerHomeIpPortsPreview from './ContainerHomeIpPortsPreview.react';
 
 var ContainerHomePreview = React.createClass({
   contextTypes: {
@@ -51,41 +51,7 @@ var ContainerHomePreview = React.createClass({
     if (this.props.defaultPort) {
       preview = (<ContainerHomeWebPreview ports={this.props.ports} defaultPort={this.props.defaultPort}/>);
     } else {
-      var ports = _.map(_.pairs(this.props.ports), pair => {
-        var key = pair[0];
-        var val = pair[1];
-        return (
-          <tr key={key}>
-            <td>{key + '/' + val.portType}</td>
-            <td>{val.url}</td>
-          </tr>
-        );
-      });
-
-      preview = (
-        <div className="web-preview wrapper">
-          <div className="widget">
-            <div className="top-bar">
-              <div className="text">IP & PORTS</div>
-              <div className="action" onClick={this.handleClickNotShowingCorrectly}>
-                <span className="icon icon-preferences"></span>
-              </div>
-            </div>
-            <p>You can access this container using the following IP address and port:</p>
-            <table className="table">
-              <thead>
-                <tr>
-                  <th>DOCKER PORT</th>
-                  <th>ACCESS URL</th>
-                </tr>
-              </thead>
-              <tbody>
-                {ports}
-              </tbody>
-            </table>
-          </div>
-        </div>
-      );
+      preview = (<ContainerHomeIpPortsPreview ports={this.props.ports}/>);
     }
     return preview;
   }
